+++ TrainTry/appsettings.json
@@ -1,6 +1,6 @@
 {
 	"ConnectionStrings": {
-		"DefaultConnection": "Host=localhost;Port=5432;Database=TrainTry;Username=postgres;Password=123"
+		"DefaultConnection": "Host=localhost;Port=5432;Database=TrainTry;Username=postgres;Password=9001"
 	},
