+++ TryTrain_Tests/UnitTests.cs
@@ -1,0 +1,127 @@
+﻿using Microsoft.Playwright;
+using Microsoft.VisualStudio.TestPlatform.ObjectModel.Client;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Text.Json;
+using System.Threading.Tasks;
+
+namespace UnitTests
+{
+    internal class UnitTestsMemorableDates
+    {
+        private IPlaywright _playwright;
+        private IAPIRequestContext _request;
+
+        [SetUp]
+        public async Task Setup()
+        {
+            _playwright = await Playwright.CreateAsync();
+
+            _request = await _playwright.APIRequest.NewContextAsync(new APIRequestNewContextOptions
+            {
+                BaseURL = "https://localhost:7149/",
+                IgnoreHTTPSErrors = true
+            });
+        }
+
+        #region [Удаление нулевого id]
+        
+        [Test]
+        public async Task DeleteMemorableDateTest1()
+        {
+            var token = await GetToken();
+
+            var response = await _request.DeleteAsync(url: "MemorableDates/DeleteMemorablesDates", new APIRequestContextOptions()
+            {
+                Headers = new Dictionary<string, string>
+                {
+                    { "Authorization", $"Bearer {token}" }
+                },
+                DataObject = new Dictionary<string, object>
+                {
+                    { "id", "0" }
+                }
+            });
+
+            Assert.AreEqual(404, response.Status);
+        }
+
+        #endregion
+
+        #region [Тест удаления id меньше нуля]
+
+        [Test]
+        public async Task DeleteMemorableDateTest2()
+        {
+            var token = await GetToken();
+
+            var response = await _request.DeleteAsync(url: "MemorableDates/DeleteMemorablesDates", new APIRequestContextOptions()
+            {
+                Headers = new Dictionary<string, string>
+                {
+                    { "Authorization", $"Bearer {token}" }
+                },
+                DataObject = new Dictionary<string, object>
+                {
+                    { "id", "-1" }
+                }
+            });
+
+            Assert.AreEqual(404, response.Status);
+        }
+
+        #endregion
+
+        #region [Удаление id с слишком большим числом]
+
+        [Test]
+        public async Task DeleteMemorableDateTest3()
+        {
+            var token = await GetToken();
+
+            var response = await _request.DeleteAsync(url: "MemorableDates/DeleteMemorablesDates", new APIRequestContextOptions()
+            {
+                Headers = new Dictionary<string, string>
+                {
+                    { "Authorization", $"Bearer {token}" }
+                },
+                DataObject = new Dictionary<string, object>
+                {
+                    { "id", "999999999999999999999999999999" }
+                }
+            });
+
+            Assert.AreEqual(404, response.Status);
+        }
+
+        #endregion
+
+        #region [Таск получения токена для тестов]
+
+        private async Task<string?> GetToken()
+        {
+
+            var response = await _request.PostAsync(url: "Account/login", new APIRequestContextOptions()
+            {
+                DataObject = new
+                {
+                    login = "Neralis",
+                    password = "9001"
+                }
+            });
+
+            var responseBody = await response.TextAsync();
+
+            // Десериализация JSON тела ответа в объект (предполагаем, что токен находится в поле 'token')
+            var responseObject = JsonSerializer.Deserialize<Dictionary<string, object>>(responseBody);
+
+            var token = responseObject["token"].ToString();
+
+            return token;
+        }
+
+        #endregion
+    }
+}
